+++ applications/content/src/main/java/org/apache/ofbiz/content/survey/PdfSurveyServices.java
@@ -152,10 +152,10 @@ public class PdfSurveyServices {
 
                 Long sequenceNum = null;
                 if (tabPage != null && tabOrder != null) {
-                    sequenceNum = Long.valueOf(tabPage.intValue() * 1000 + tabOrder.intValue());
+                    sequenceNum = (long) (tabPage * 1000 + tabOrder);
                     Debug.logInfo("tabPage=" + tabPage + ", tabOrder=" + tabOrder + ", sequenceNum=" + sequenceNum, module);
                 } else if (fieldPositions.length > 0) {
-                    sequenceNum = Long.valueOf((long) fieldPage * 10000 + (long) fieldLly * 1000 + (long) fieldLlx);
+                    sequenceNum = (long) fieldPage * 10000 + (long) fieldLly * 1000 + (long) fieldLlx;
                     Debug.logInfo("fieldPage=" + fieldPage + ", fieldLlx=" + fieldLlx + ", fieldLly=" + fieldLly + ", fieldUrx=" + fieldUrx + ", fieldUry=" + fieldUry + ", sequenceNum=" + sequenceNum, module);
                 }
