+++ applications/marketing/src/main/java/org/apache/ofbiz/marketing/report/ReportHelper.java
@@ -64,21 +64,21 @@ public final class ReportHelper {
 
                 reportValue.put("orders", orderValue.getLong("orderId")); // # of orders
                 if (orderValue.getDouble("grandTotal") == null) {
-                    reportValue.put("orderAmount", Double.valueOf(0));
+                    reportValue.put("orderAmount", (double) 0);
                 } else {
                     reportValue.put("orderAmount", orderValue.getDouble("grandTotal"));
                 }
                 if ((orderValue.getLong("orderId") == null) || (visit.getLong("visitId") == null) ||
                     (visit.getLong("visitId").intValue() == 0)) {
-                    reportValue.put("conversionRate", Double.valueOf(0));
+                    reportValue.put("conversionRate", (double) 0);
                 } else {
-                    reportValue.put("conversionRate", Double.valueOf(orderValue.getLong("orderId").doubleValue() / visit.getLong("visitId").doubleValue()));
+                    reportValue.put("conversionRate", orderValue.getLong("orderId").doubleValue() / visit.getLong("visitId").doubleValue());
                 }
             } else {
                 // no matching orders - all those values are zeroes
-                reportValue.put("orders", Long.valueOf(0));
-                reportValue.put("orderAmount", Double.valueOf(0));
-                reportValue.put("conversionRate", Double.valueOf(0));
+                reportValue.put("orders", 0L);
+                reportValue.put("orderAmount", (double) 0);
+                reportValue.put("conversionRate", (double) 0);
             }
 
             conversionRates.add(reportValue);
