+++ framework/entity/src/main/java/org/apache/ofbiz/entity/jdbc/CursorStatement.java
@@ -81,9 +81,9 @@ public class CursorStatement extends AbstractCursorHandler {
         } else if ("setCursorName".equals(method.getName())) {
             setCursorName((String) args[0]);
         } else if ("getFetchSize".equals(method.getName())) {
-            return Integer.valueOf(getFetchSize());
+            return getFetchSize();
         } else if ("setFetchSize".equals(method.getName())) {
-            setFetchSize(((Integer) args[0]).intValue());
+            setFetchSize((Integer) args[0]);
         }
         return super.invoke(stmt, proxy, method, args);
     }
